Updated freeimage to 3.17.0. This commit has been slightly modified from the original freeimage source code in that it has had some guards added to a few files to work around snprintf on vs 2015, hopefully this will be fixed in the next version.
We are also building without LibRaw since it seems to cause a lot of problems for us on other platforms. To fix this a file was added called PluginFakeRAW.

This version needs to be cleaned, it contains more files than we need, but this gets it merged and it works on windows at least. This commit should be considered a reset of any modifications we have made to freeimage in the past.